requirements.txt
Flask==2.3.3openai

Werkzeug==2.3.7
python-docx==0.8.11
PyPDF2==3.0.1
openai==1.3.5
python-dotenv==1.0.0
gunicorn==21.2.0
google-generativeai==0.3.2
requests==2.31.0
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0

# Additional packages for enhanced system management and diagnostics
psutil==5.9.6
subprocess32==3.5.4; python_version < '3.3'

# Enhanced error handling and debugging
traceback2==1.4.0

# Improved file handling and encoding support
chardet==5.2.0
pathlib2==2.3.7; python_version < '3.4'

# Better subprocess and system interaction (for build diagnostics)
sh==2.0.6
Flask-Login==0.6.3

# Note: pdflatex CANNOT be installed via pip!
# It's a system-level LaTeX compiler that must be installed via:
# - System package manager (apt, yum, etc.)
# - Manual installation (TeX Live)
# - Our comprehensive build.sh script handles this
# 
# The build.sh script uses "MAXIMUM FORCE" installation strategies:
# 1. Multiple apt-get attempts with different flags
# 2. Individual package installation with fallbacks
# 3. Manual TeX Live installation from source
# 4. Symlink creation and PATH management
# 5. Enhanced wrapper scripts as final fallback
